--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,6 @@
 module github.com/ua-parser/uap-go
 
-require gopkg.in/yaml.v2 v2.2.1
+require (
+	github.com/hashicorp/golang-lru v0.5.4
+	gopkg.in/yaml.v2 v2.2.1
+)