The caching is a bit janky and atypical of a parsing library, but I think the performance improvements justify it.

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkParser-32                13418192      38386         -99.71%
BenchmarkParserWithOptions-32     10629612      24925         -99.77%
```