@dgoldstein0 2023-08-23
I agree that caching is a good idea - it's very common to use this library e.g. in a serving use case, where the most common user agent strings make up almost all calls.  and other uap libraries do the same - the one I'm familiar with is uap-python.

Out of curiosity why do you call this "janky"?