@ammario 2023-08-23
Hey, happy to see a response here :)

It's just a little bit janky in that it's rare for parsing libraries to use caching like this, but the jank is definitely worth it.

We've been using the proposed changes to great success in https://github.com/coder/coder for over a year. UAP no longer shows up on our flame graphs.